% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ie.r
\name{ie}
\alias{ie}
\title{Function for the method of Independent Evolution}
\usage{
ie(data, tree)
}
\arguments{
\item{data}{numeric vector with elements in the same order as tiplabels in the tree}
}
\value{
dataframe with rescaled branch lengths (rBL) for all branches in the tree
}
\description{
This function allows computing rescaled branch lengths according to inferred phenotypic evolution and ancestral states using the method of Independent Evolution as described in Smaers & Vinicius (2009)
}
\details{
This function poses the following restrictions on the input data: no polytomies in the tree, no duplicated values, no zero values, and no negative values in the data. The algorithm was designed to deal with linear data (unlogged).
}

